+++ packages/cli/src/utils/ConsoleReporter.ts
@@ -92,8 +92,10 @@ cause.stack: ${detail.causeStack}`)
 
 	testScriptConsole(method: string, message?: any, ...optionalParams: any[]): void {
 		debug('testScriptConsole', method, message)
+		if (method === 'clear') return
 		if (method === 'log') method = 'info'
 		if (method === 'warning') method = 'warn'
-		;(this.logger as any)[method](`page console.${method}: ${message} ${optionalParams.join(' ')}`)
+		const consolMethod = this.logger[method] || console.info
+		consolMethod(`page console.${method}: ${message} ${optionalParams.join(' ')}`)
 	}
 }