+++ packages/docs/CHANGELOG.md
@@ -17,7 +17,41 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.3.0-beta.28](https://github.com/flood-io/element/compare/v1.3.0-beta.27...v1.3.0-beta.28) (2020-08-29)
 
+**Note:** Version bump only for package element-docs
+
+
+
+
+
+## [1.3.5](https://github.com/flood-io/element/compare/v1.3.4...v1.3.5) (2020-09-22)
+
+**Note:** Version bump only for package element-docs
+
+
+
+
+
+## [1.3.4](https://github.com/flood-io/element/compare/v1.3.3...v1.3.4) (2020-09-18)
+
+**Note:** Version bump only for package element-docs
+
+
+
+
+
+## [1.3.1](https://github.com/flood-io/element/compare/v1.3.0-beta.28...v1.3.1) (2020-09-04)
+
+
+### Bug Fixes
+
+* bump 1.3.1 ([#262](https://github.com/flood-io/element/issues/262)) ([b63ed9a](https://github.com/flood-io/element/commit/b63ed9a0d27e1184e377e7f28852fb70a1895d73))
+
+
+
+
+
+# [1.3.0](https://github.com/flood-io/element/compare/v1.3.0-beta.20...v1.3.0) (2020-08-29)
 
 ### Features
 
-* new documentation website ([#178](https://github.com/flood-io/element/issues/178)) ([2b13b14](https://github.com/flood-io/element/commit/2b13b142b7f347b2f9dc1a67334bdc69ef5fff69)), closes [#187](https://github.com/flood-io/element/issues/187) [#232](https://github.com/flood-io/element/issues/232)
+- new documentation website ([#178](https://github.com/flood-io/element/issues/178)) ([2b13b14](https://github.com/flood-io/element/commit/2b13b142b7f347b2f9dc1a67334bdc69ef5fff69)), closes [#187](https://github.com/flood-io/element/issues/187) [#232](https://github.com/flood-io/element/issues/232)